add react-router-dom and created Favoirte page
