+++ src/App.js
@@ -1,54 +1,37 @@
-import React, { useEffect } from "react";
+import React from "react";
 import "./App.css";
+import Home from "./pages/Home";
+import {
+  BrowserRouter as Router,
+  Route,
+  Switch,
+  Redirect,
+  Link,
+} from "react-router-dom";
+import Favorite from "./pages/Favorite";
 import styled from "styled-components";
 import LogoImg from "./images/Logo.svg";
-import BubbleOne from "./images/BubbleOne.svg";
-import BubbleTwo from "./images/BubbleTwo.svg";
-import SearchBar from "./components/SearchBar";
-import TempCards from "./components/TempCards";
-import axios from "axios";
-import { setWeather } from "./features/weatherSlice";
-import { useDispatch } from "react-redux";
 
 function App() {
-  const dispatch = useDispatch();
-  useEffect(() => {
-    const fetchData = async () => {
-      await axios
-        .get(
-          "http://api.openweathermap.org/data/2.5/weather?q=tel-aviv&appid=b0a3a56cdf847e01bafd3230c4409e71&units=metric"
-        )
-        .then(({ data }) => {
-          // console.log(data);
-          dispatch(
-            setWeather({
-              city: data.name,
-              temp: data.main.temp,
-              description: data.weather[0].description,
-            })
-          );
-        });
-    };
-    fetchData();
-  }, [dispatch]);
   return (
-    <Container>
-      <LogoContainer>
-        <ImageContiner>
-          <ImageLogo src={LogoImg} alt="Logo" />
-        </ImageContiner>
-      </LogoContainer>
-      <BubbleTop src={BubbleOne} alt="Bubble" />
-      <BubbleBottom src={BubbleTwo} alt="Bubble" />
-      <GlassContainer>
-        <Glass>
-          <SearchBar />
-          <CardContainer>
-            <TempCards />
-          </CardContainer>
-        </Glass>
-      </GlassContainer>
-    </Container>
+    <Router>
+      <>
+        <Container>
+          <LogoContainer>
+            <ImageContiner>
+              <Link to="/">
+                <ImageLogo src={LogoImg} alt="Logo" />
+              </Link>
+            </ImageContiner>
+          </LogoContainer>
+          <Switch>
+            <Route exact path="/" component={Home} />
+            <Route exact path="/favorite" component={Favorite} />
+            <Redirect to="/" />
+          </Switch>
+        </Container>
+      </>
+    </Router>
   );
 }
 
@@ -73,43 +56,6 @@ const ImageContiner = styled.div`
 const ImageLogo = styled.img`
   width: 100%;
   height: 100%;
-`;
-const BubbleTop = styled.img`
-  position: absolute;
-  top: -150px;
-  right: -180px;
-  height: auto;
-  width: 45rem;
-`;
-const BubbleBottom = styled.img`
-  position: absolute;
-  bottom: -200px;
-  left: 20px;
-  height: auto;
-  width: 30rem;
-`;
-
-const GlassContainer = styled.div`
-  margin-top: 1.2rem;
-  width: 100vw;
-  min-height: 800px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-const Glass = styled.div`
-  height: 745px;
-  width: 75%;
-  background: rgba(255, 255, 255, 0.2);
-  border-radius: 10px;
-  backdrop-filter: blur(10px);
-`;
-
-const CardContainer = styled.div`
-  margin-top: 50px;
-  display: flex;
-  justify-content: center;
 `;
 
 export default App;
