+++ src/pages/Favorite.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Favorite = () => {
+  return <div>Favorite!!!</div>;
+};
+
+export default Favorite;
